+++ src/App.jsx
@@ -1,12 +1,13 @@
 // src/App.jsx
 import React from 'react';
 import AppRoutes from './routes/AppRoutes';
+// import Home from './pages/Home';
 
 
 const App = () => {
     return (
         <div className="App">
-            <AppRoutes />
+             <AppRoutes /> 
         </div>
     );
 };
