+++ src/components/NavBar/NavBar.css
@@ -1,5 +1,7 @@
 .navbar { 
+    display: flex;
     padding: 1rem; display: flex; 
+    align-items: center;
     justify-content: space-between; 
     align-items: center; 
     color: #ffffff; 
@@ -9,6 +11,8 @@
 .container-fluid { 
     width: 100%; 
     padding: 0 2rem; 
+    display: flex;
+    align-items: center;
 }
 
 .navbar-brand { 
@@ -19,4 +23,10 @@
 }
 .navbar-brand img{
     margin-right: 1rem;
+}
+.navbar-children{
+    margin-left: auto;
+    display: flex;
+    align-items: center;
+    gap: 2rem;
 }
