+++ src/main.jsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
 import { AuthProvider } from './contexts/AuthContext';
-import './styles/Main.css';
+import './styles/Global.css';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
