--- src/styles/Main.css
@@ -1,5 +1,0 @@
-body{
-    margin: 0;
-    font-family: "Work Sans", sans-serif;
-    background-color: #F7FAFC;
-}
