+++ src/styles/global.css
@@ -1,112 +1,21 @@
-/* src/styles/global.css
+.container{
+    display: flex;
+    align-items: center;
+    height: 100vh;
+    flex-direction: column;
+    margin: 0 auto;
+    text-align: center;
+    margin-top: 3rem;
+    width: 70%;
+}
+
 body{
     margin: 0;
-    padding: 0;
+    font-family: "Work Sans", sans-serif;
+    background-color: #F7FAFC;
 }
-.idea-generator-form {
-    max-width: 600px;
-    margin: 20px auto;
-    padding: 20px;
-    background: #fff;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+a{
+    text-decoration: none;
+    color: #0D141C;
+    font-weight: 500;
 }
-
-h1 {
-    text-align: center;
-    color: #333;
-}
-
-form {
-    display: flex;
-    flex-direction: column;
-}
-
-label {
-    margin-bottom: 15px;
-}
-
-input, select {
-    padding: 10px;
-    margin-top: 5px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-}
-
-button {
-    padding: 10px;
-    background-color: #007BFF;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #0056b3;
-}
-
-
-.idea-generator-form {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    background: #f9f9f9;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-h1 {
-    text-align: center;
-}
-
-label {
-    display: block;
-    margin: 15px 0 5px;
-}
-
-fieldset {
-    margin: 10px 0;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    padding: 10px;
-}
-
-legend {
-    font-weight: bold;
-}
-
-input[type="text"],
-select {
-    width: 100%;
-    padding: 8px;
-    margin-top: 5px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-}
-
-button {
-    margin-top: 10px;
-    padding: 10px;
-    width: 100%;
-    background-color: #4CAF50;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #45a049;
-}
-
-.checkbox-grid {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr); /* 4 columnas de igual ancho */
-    /* gap: 10px; /* Espaciado entre las casillas de verificación */
-/* }
-
-.checkbox-grid label {
-    display: flex;
-    align-items: center; /* Alinear verticalmente 
-} */
